DDF-1479 Update solr from 4.7.2 to 4.10.4
DDF-1479 - Fixed some minor errors in container registration

Because of lux, you currently have to instantiate the SolrCore and pass
in the schema as a parameter, because lux will look for it on creation.
If it is not there (eg, setting it after creation) then it will not hook
in the correct fields and will fail.

DDF-1479 Change all LuxDB versions to 1.1.1-SNAPSHOT

DDF-1479 Fix to testSolrBackUpNumToKeep() because new Solr version changed how -n flag works

DDF-1479 Update FilterSolrIndex to use new implementation for SolrCore creation

DDF-1479 Add lucene-join in the embedded dependency section of a couple poms within the Catalog

DDF-1479 Extracted new SolrCoreContainer to its own class