DDF-1476 - Updating bounding box causes inconsistent behavior
When using the numeric input boxes to adjust the bounding box
North and South values, the box is either removed or moved to a
unknown location on the 2d map. This was because the values are
not being treated as numeric. This was solved by ensuring they are
numeric once they are retrieved from model.