DDF-1493 Remove gazetteer feature from search-ui-app feature.
Prior to this change, the webservice-gazetteer was being loaded as part of the
search-ui feature.  While the search-ui needs a gazetteer, the selection of a
specific gazetteer implementation should be left to configurations that are
loaded at a later time.  This change adds a ServiceSelector which determines
which GeoCoder implementation should be used based on the service ranking of
each bound implementation of GeoCoder.  It also adds new unit tests for the
ServiceSelector and the GeoCoderEndpoint.