DDF-4478 Improves filtering logic for dropdowns (#4254)
Improves filtering logic for dropdowns in Intrigue.
Previously, filtering through a dropdown would match if any part of the filter was found in the word. This would lead to weird behavior like returning "validation" and "width" when the filter was "id".
The new behavior splits the words on camelcase, dashes, dots, and spaces then checks to see if any of the resulting words start with the filter value. This results in a slightly more intuitive behavior.