DDF-5649 G-6176 G-7742 Show warnings to the user
This commit enables that DDF's UI display to the user any warning about
a malfunction of a source which a filter has deemed relevant and useful.
With this enhancement, any source can convey to the user comprehensible
information about the cause of a malfunction or how the user might
resolve the malfunction.