Add 'featured attributes' configuration
Admins can now set a list of 'featured attributes', which will be placed
in a distinct group at the top of the drop-down list displayed when
choosing an attribute in the advanced search menu. The complete list of
attributes, which includes the featured attributes, will be displayed in
a separate group below.