Improve location input error handling (#756)
This commit enhances the location inputs, which now accept a prop (errorListener) that can be used to notify interested parties of validation errors. It also eliminates the use of onBlur for validation, so users can get immediate feedback as they type. Finally, it makes several improvements to location validation to handle more cases.